Rename PolarisMetaStoreManager and its sibling interfaces
