+++ polaris-core/src/main/java/org/apache/polaris/core/storage/cache/StorageCredentialCache.java
@@ -34,8 +34,8 @@ import org.apache.polaris.core.config.FeatureConfiguration;
 import org.apache.polaris.core.config.PolarisConfiguration;
 import org.apache.polaris.core.entity.PolarisEntity;
 import org.apache.polaris.core.entity.PolarisEntityType;
+import org.apache.polaris.core.persistence.dao.CredentialVendorDao;
 import org.apache.polaris.core.persistence.dao.entity.ScopedCredentialsResult;
-import org.apache.polaris.core.storage.PolarisCredentialVendor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -100,7 +100,7 @@ public class StorageCredentialCache {
    * @return the a map of string containing the scoped creds information
    */
   public Map<String, String> getOrGenerateSubScopeCreds(
-      @Nonnull PolarisCredentialVendor credentialVendor,
+      @Nonnull CredentialVendorDao credentialVendor,
       @Nonnull PolarisCallContext callCtx,
       @Nonnull PolarisEntity polarisEntity,
       boolean allowListOperation,
